# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FE UCA Facturacion Electronica Backend
  version: 0.1.0
  description: Especificacion inicial de endpoints principales. Las respuestas usan el envoltorio GeneralResponse.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/empresas/{empresaId}/facturas:
    get:
      summary: Listar facturas.
      parameters:
        - $ref: '#/components/parameters/empresaId'
      responses:
        '200':
          description: Facturas encontradas.
    post:
      summary: Crear factura en BORRADOR.
      parameters:
        - $ref: '#/components/parameters/empresaId'
      responses:
        '201':
          description: Factura creada.
  /api/v1/empresas/{empresaId}/facturas/{id}:
    get:
      summary: Consultar factura.
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Factura encontrada.
    put:
      summary: Actualizar factura en BORRADOR.
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Factura actualizada.
    delete:
      summary: Eliminar factura editable.
      description: No elimina facturas emitidas.
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Factura eliminada.
  /api/v1/empresas/{empresaId}/facturas/{id}/preparar-envio:
    post:
      summary: Validar y preparar DTE para envio.
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: DTE listo para envio.
  /api/v1/empresas/{empresaId}/facturas/{id}/enviar:
    post:
      summary: Enviar factura a Hacienda.
      description: Requiere ADMINISTRADOR o SUPERADMIN. Soporta Idempotency-Key.
      parameters:
        - $ref: '#/components/parameters/empresaId'
        - $ref: '#/components/parameters/id'
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Respuesta de Hacienda procesada.
        '409':
          description: Conflicto de estado o idempotencia.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    empresaId:
      name: empresaId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...